requirements.txt
openai
fastapi
uvicorn
Pillow
python-multipart
